// hooks/api/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Any: any;
  timestamptz: any;
  uuid: any;
};
// ...
export type Frequency = {
  __typename?: 'frequency';
  currency?: Maybe<Scalars['String']>;
  description: Scalars['String'];
  montly_price: Scalars['String'];
  /** An array relationship */
  smartsheets: Array<Smartsheet>;
  /** An aggregate relationship */
  smartsheets_aggregate: Smartsheet_Aggregate;
  title: Scalars['String'];
  yearly_price: Scalars['String'];
};
// ...
export type Smartsheet = {
  __typename?: 'smartsheet';
  createdAt?: Maybe<Scalars['timestamptz']>;
  frequency?: Maybe<Scalars['String']>;
  /** An object relationship */
  frequencyByFrequency?: Maybe<Frequency>;
  id: Scalars['uuid'];
  is_monthly_payment?: Maybe<Scalars['Boolean']>;
  is_payment_set_up?: Maybe<Scalars['Boolean']>;
  lastSyncedAt?: Maybe<Scalars['timestamptz']>;
  rules_sheet?: Maybe<Scalars['String']>;
  smart_sheet_api_token: Scalars['String'];
  updatedAt?: Maybe<Scalars['timestamptz']>;
  /** An object relationship */
  user: User;
  userId: Scalars['uuid'];
};

/** aggregated selection of "smartsheet" */
export type Smartsheet_Aggregate = {
  __typename?: 'smartsheet_aggregate';
  aggregate?: Maybe<Smartsheet_Aggregate_Fields>;
  nodes: Array<Smartsheet>;
};
// ...
export type Smartsheet_Aggregate_Fields = {
  __typename?: 'smartsheet_aggregate_fields';
  count: Scalars['Int'];
  max?: Maybe<Smartsheet_Max_Fields>;
  min?: Maybe<Smartsheet_Min_Fields>;
};
// ...
export type User = {
  __typename?: 'user';
  createdAt: Scalars['timestamptz'];
  email: Scalars['String'];
  firstName?: Maybe<Scalars['String']>;
  id: Scalars['uuid'];
  isActive: Scalars['Boolean'];
  lastName: Scalars['String'];
  password?: Maybe<Scalars['String']>;
  profileImage?: Maybe<Scalars['String']>;
  provider?: Maybe<Scalars['String']>;
  smart_sheet_mgt_is_active?: Maybe<Scalars['Boolean']>;
  /** An array relationship */
  smartsheets: Array<Smartsheet>;
  /** An aggregate relationship */
  smartsheets_aggregate: Smartsheet_Aggregate;
  updatedAt: Scalars['timestamptz'];
};